Dummy classes complete

+++ GesTur/app/src/main/java/com/example/gestur/logic/questions/ScoreObsQuestion.java
@@ -1,10 +1,12 @@
 package com.example.gestur.logic.questions;
 
+import com.example.gestur.DB.definitionFormClasses.QuestionDefScoreObs;
+
 public class ScoreObsQuestion extends ScoreQuestion {
 
     private String observation;
 
-    public ScoreObsQuestion(String number, short min, short max, String question, Boolean NA, short score, String observation) {
+    public ScoreObsQuestion(String number, int min, int max, String question, Boolean NA, int score, String observation) {
         super(number, min, max, question, NA, score);
         this.observation = observation;
     }
